wip - adding other procs
