+++ extended-it/src/test/java/apoc/s3/ArrowS3Test.java
@@ -1,0 +1,55 @@
+package apoc.s3;
+
+import apoc.export.arrow.ArrowTestUtil;
+import apoc.util.s3.S3BaseTest;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.neo4j.configuration.GraphDatabaseInternalSettings;
+import org.neo4j.test.rule.DbmsRule;
+import org.neo4j.test.rule.ImpermanentDbmsRule;
+
+import java.util.Map;
+
+import static apoc.export.arrow.ArrowTestUtil.beforeClassCommon;
+import static apoc.export.arrow.ArrowTestUtil.testImportCommon;
+import static apoc.export.arrow.ArrowTestUtil.testLoadArrow;
+
+public class ArrowS3Test extends S3BaseTest {
+
+    @Rule
+    public DbmsRule db = new ImpermanentDbmsRule()
+            .withSetting(GraphDatabaseInternalSettings.enable_experimental_cypher_versions, true);
+
+    @Before
+    public void beforeClass() {
+        beforeClassCommon(db);
+    }
+
+    @Test
+    public void testFileRoundtripWithLoadArrow() {
+        String url = s3Container.getUrl("test_all.arrow");
+        
+        String file = db.executeTransactionally("CYPHER 25 CALL apoc.export.arrow.all($url) YIELD file",
+                Map.of("url", url),
+                ArrowTestUtil::extractFileName);
+
+        // check that the exported file is correct
+        final String query = "CYPHER 25 CALL apoc.load.arrow($file, {})";
+        testLoadArrow(db, query, Map.of("file", file));
+    }
+
+
+    @Test
+    public void testFileRoundtripWithImportArrow() {
+        db.executeTransactionally("CREATE (:Another {foo:1, listInt: [1,2]}), (:Another {bar:'Sam'})");
+
+        String url = s3Container.getUrl("test_all_import.arrow");
+        String file = db.executeTransactionally("CYPHER 25 CALL apoc.export.arrow.all($url) YIELD file",
+                Map.of("url", url),
+                ArrowTestUtil::extractFileName);
+        
+        // check that the exported file is correct
+        testImportCommon(db, file, ArrowTestUtil.MAPPING_ALL);
+    }
+}
