+++ extended/src/test/java/apoc/export/arrow/ArrowExtendedTest.java
@@ -31,6 +31,7 @@ import org.junit.ClassRule;
 import org.junit.Test;
 import org.neo4j.configuration.GraphDatabaseInternalSettings;
 import org.neo4j.configuration.GraphDatabaseSettings;
+import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Result;
 import org.neo4j.test.rule.DbmsRule;
 import org.neo4j.test.rule.ImpermanentDbmsRule;
@@ -48,6 +49,8 @@ import java.util.stream.LongStream;
 import static apoc.ApocConfig.APOC_EXPORT_FILE_ENABLED;
 import static apoc.ApocConfig.APOC_IMPORT_FILE_ENABLED;
 import static apoc.ApocConfig.apocConfig;
+import static apoc.export.arrow.ArrowTestUtil.beforeClassCommon;
+import static apoc.export.arrow.ArrowTestUtil.testLoadArrow;
 import static org.junit.Assert.assertEquals;
 
 public class ArrowExtendedTest {
@@ -65,70 +68,11 @@ public class ArrowExtendedTest {
                     directory.toPath().toAbsolutePath())
             .withSetting(GraphDatabaseInternalSettings.enable_experimental_cypher_versions, true);
 
-    public static final List<Map<String, Object>> EXPECTED = List.of(
-            new HashMap<>() {
-                {
-                    put("name", "Adam");
-                    put("bffSince", null);
-                    put("<source.id>", null);
-                    put("<id>", 0L);
-                    put("age", 42L);
-                    put("labels", List.of("User"));
-                    put("male", true);
-                    put("<type>", null);
-                    put("kids", List.of("Sam", "Anna", "Grace"));
-                    put(
-                            "place",
-                            Map.of("crs", "wgs-84-3d", "longitude", 33.46789D, "latitude", 13.1D, "height", 100.0D));
-                    put("<target.id>", null);
-                    put("since", null);
-                    put(
-                            "born",
-                            LocalDateTime.parse("2015-05-18T19:32:24.000")
-                                    .atOffset(ZoneOffset.UTC)
-                                    .toZonedDateTime());
-                }
-            },
-            new HashMap<>() {
-                {
-                    put("name", "Jim");
-                    put("bffSince", null);
-                    put("<source.id>", null);
-                    put("<id>", 1L);
-                    put("age", 42L);
-                    put("labels", List.of("User"));
-                    put("male", null);
-                    put("<type>", null);
-                    put("kids", null);
-                    put("place", null);
-                    put("<target.id>", null);
-                    put("since", null);
-                    put("born", null);
-                }
-            },
-            new HashMap<>() {
-                {
-                    put("name", null);
-                    put("bffSince", "P5M1DT12H");
-                    put("<source.id>", 0L);
-                    put("<id>", 0L);
-                    put("age", null);
-                    put("labels", null);
-                    put("male", null);
-                    put("<type>", "KNOWS");
-                    put("kids", null);
-                    put("place", null);
-                    put("<target.id>", 1L);
-                    put("since", 1993L);
-                    put("born", null);
-                }
-            });
+
 
     @BeforeClass
     public static void beforeClass() {
-        db.executeTransactionally(
-                "CREATE (f:User {name:'Adam',age:42,male:true,kids:['Sam','Anna','Grace'], born:localdatetime('2015-05-18T19:32:24.000'), place:point({latitude: 13.1, longitude: 33.46789, height: 100.0})})-[:KNOWS {since: 1993, bffSince: duration('P5M1.5D')}]->(b:User {name:'Jim',age:42})");
-        TestUtil.registerProcedure(db, ExportArrowExtended.class, LoadArrowExtended.class, Graphs.class, Meta.class);
+        beforeClassCommon(db);
     }
 
     @AfterClass
@@ -136,27 +80,8 @@ public class ArrowExtendedTest {
         db.shutdown();
     }
 
-    @Before
-    public void before() {
-        apocConfig().setProperty(APOC_IMPORT_FILE_ENABLED, true);
-        apocConfig().setProperty(APOC_EXPORT_FILE_ENABLED, true);
-    }
-
     private byte[] extractByteArray(Result result) {
         return result.<byte[]>columnAs("byteArray").next();
-    }
-
-    private String extractFileName(Result result) {
-        return result.<String>columnAs("file").next();
-    }
-
-    private <T> T readValue(String json, Class<T> clazz) {
-        if (json == null) return null;
-        try {
-            return JsonUtil.OBJECT_MAPPER.readValue(json, clazz);
-        } catch (JsonProcessingException e) {
-            return null;
-        }
     }
 
     @Test
@@ -215,7 +140,7 @@ public class ArrowExtendedTest {
         String file = db.executeTransactionally(
                 "CYPHER 25 CALL apoc.export.arrow.query('query_test.arrow', $query) YIELD file",
                 Map.of("query", returnQuery),
-                this::extractFileName);
+                ArrowTestUtil::extractFileName);
 
         // then
         final String query = "CYPHER 25 CALL apoc.load.arrow($file) YIELD value " + "RETURN value";
@@ -251,22 +176,7 @@ public class ArrowExtendedTest {
 
         // then
         final String query = "CYPHER 25 CALL apoc.load.arrow.stream($byteArray) YIELD value " + "RETURN value";
-        db.executeTransactionally(query, Map.of("byteArray", byteArray), result -> {
-            final List<Map<String, Object>> actual = getActual(result);
-            assertEquals(EXPECTED, actual);
-            return null;
-        });
-    }
-
-    private List<Map<String, Object>> getActual(Result result) {
-        return result.stream()
-                .map(m -> (Map<String, Object>) m.get("value"))
-                .map(m -> {
-                    final Map<String, Object> newMap = new HashMap(m);
-                    newMap.put("place", readValue((String) m.get("place"), Map.class));
-                    return newMap;
-                })
-                .collect(Collectors.toList());
+        testLoadArrow(db, query, Map.of("byteArray", byteArray));
     }
 
     @Test
@@ -277,15 +187,11 @@ public class ArrowExtendedTest {
                         + "CALL apoc.export.arrow.graph('graph_test.arrow', graph) YIELD file "
                         + "RETURN file",
                 Map.of(),
-                this::extractFileName);
+                ArrowTestUtil::extractFileName);
 
         // then
         final String query = "CYPHER 25 CALL apoc.load.arrow($file) YIELD value " + "RETURN value";
-        db.executeTransactionally(query, Map.of("file", file), result -> {
-            final List<Map<String, Object>> actual = getActual(result);
-            assertEquals(EXPECTED, actual);
-            return null;
-        });
+        testLoadArrow(db, query, Map.of("file", file));
     }
 
     @Test
@@ -310,26 +216,18 @@ public class ArrowExtendedTest {
 
         // then
         final String query = "CYPHER 25 CALL apoc.load.arrow.stream($byteArray) YIELD value " + "RETURN value";
-        db.executeTransactionally(query, Map.of("byteArray", byteArray), result -> {
-            final List<Map<String, Object>> actual = getActual(result);
-            assertEquals(EXPECTED, actual);
-            return null;
-        });
+        testLoadArrow(db, query, Map.of("byteArray", byteArray));
     }
 
     @Test
     public void testFileRoundtripArrowAll() {
         // given - when
         String file = db.executeTransactionally(
-                "CYPHER 25 CALL apoc.export.arrow.all('all_test.arrow') YIELD file", Map.of(), this::extractFileName);
+                "CYPHER 25 CALL apoc.export.arrow.all('all_test.arrow') YIELD file", Map.of(), ArrowTestUtil::extractFileName);
 
         // then
         final String query = "CYPHER 25 CALL apoc.load.arrow($file) YIELD value " + "RETURN value";
-        db.executeTransactionally(query, Map.of("file", file), result -> {
-            final List<Map<String, Object>> actual = getActual(result);
-            assertEquals(EXPECTED, actual);
-            return null;
-        });
+        testLoadArrow(db, query, Map.of("file", file));
     }
 
     @Test
@@ -365,7 +263,7 @@ public class ArrowExtendedTest {
         String file = db.executeTransactionally(
                 "CYPHER 25 CALL apoc.export.arrow.query('volume_test.arrow', 'MATCH (n:ArrowNode) RETURN n.id AS id') YIELD file ",
                 Map.of(),
-                this::extractFileName);
+                ArrowTestUtil::extractFileName);
 
         final List<Long> expected = LongStream.range(0, 10000).mapToObj(l -> l).collect(Collectors.toList());
