+++ extended/src/test/java/apoc/export/parquet/ParquetTestUtil.java
@@ -40,7 +40,7 @@ import static org.junit.Assert.assertTrue;
 public class ParquetTestUtil {
 
     public static void beforeClassCommon(GraphDatabaseService db) {
-        TestUtil.registerProcedure(db, ExportParquet.class, ImportParquet.class, LoadParquet.class, Graphs.class, Meta.class);
+        
         
         apocConfig().setProperty(APOC_IMPORT_FILE_ENABLED, true);
         apocConfig().setProperty(APOC_EXPORT_FILE_ENABLED, true);
